moved GtSeedextendMatchIterator to own file.
