+++ src/tools/gt_show_seedext.c
@@ -28,12 +28,13 @@
 #include "match/revcompl.h"
 #include "match/ft-polish.h"
 #include "match/seed-extend.h"
-#include "match/querymatch.h"
 #include "match/seq_or_encseq.h"
+#include "match/seed-extend-iter.h"
 #include "extended/linearalign.h"
 #include "tools/gt_show_seedext.h"
 
-typedef struct {
+typedef struct
+{
   bool show_alignment,
        seed_display,
        relax_polish,
@@ -123,226 +124,14 @@ static int gt_show_seedext_arguments_check(GT_UNUSED int rest_argc,
   int had_err = 0;
 
   gt_error_check(err);
-  gt_assert(arguments);
+  gt_assert(arguments != NULL);
   if (arguments->matchfilename == NULL ||
-      gt_str_length(arguments->matchfilename) == 0) {
+      gt_str_length(arguments->matchfilename) == 0)
+  {
     gt_error_set(err,"option -f requires a file name");
     had_err = -1;
   }
   return had_err;
-}
-
-struct GtSeedextendMatchIterator
-{
-  GtStr *ii, *qii;
-  bool mirror, bias_parameters;
-  GtEncseq *aencseq, *bencseq;
-  GtUword errorpercentage, history_size, query_totallength;
-  const char *matchfilename;
-  GtStr *line_buffer;
-  uint64_t linenum;
-  bool has_seedline;
-  GtUword seedpos1,
-          seedpos2,
-          seedlen;
-  FILE *inputfileptr;
-  GtQuerymatch *querymatchptr;
-};
-
-typedef struct GtSeedextendMatchIterator GtSeedextendMatchIterator;
-
-void gt_seedextend_match_iterator_delete(GtSeedextendMatchIterator *semi)
-{
-  if (semi == NULL)
-  {
-    return;
-  }
-  gt_str_delete(semi->ii);
-  gt_str_delete(semi->qii);
-  gt_encseq_delete(semi->aencseq);
-  gt_encseq_delete(semi->bencseq);
-  gt_str_delete(semi->line_buffer);
-  gt_querymatch_delete(semi->querymatchptr);
-  if (semi->inputfileptr != NULL)
-  {
-    fclose(semi->inputfileptr);
-  }
-  gt_free(semi);
-}
-
-/* Parse encseq input indices from -ii and -qii options in 1st line of file. */
-GtSeedextendMatchIterator *gt_seedextend_match_iterator_new(
-                                            const GtStr *matchfilename,
-                                            GtError *err)
-{
-  FILE *options_line_inputfileptr;
-  int had_err = 0;
-  GtSeedextendMatchIterator *semi = gt_malloc(sizeof *semi);
-  GtEncseqLoader *encseq_loader = NULL;
-
-  semi->ii = gt_str_new(),
-  semi->qii = gt_str_new();
-  semi->mirror = false;
-  semi->bias_parameters = false;
-  semi->aencseq = semi->bencseq = NULL;
-  semi->query_totallength = 0;
-  semi->errorpercentage = 0;
-  semi->history_size = 0;
-  semi->matchfilename = gt_str_get(matchfilename);
-  semi->has_seedline = false;
-  semi->line_buffer = NULL;
-  semi->inputfileptr = NULL;
-  semi->querymatchptr = gt_querymatch_new();
-  options_line_inputfileptr = fopen(semi->matchfilename, "r");
-  if (options_line_inputfileptr == NULL) {
-    gt_error_set(err, "file %s does not exist", semi->matchfilename);
-    had_err = -1;
-  }
-  if (!had_err) {
-    GtStr *options_line_buffer = gt_str_new();
-    /* read first line and evaluate tokens */
-    if (gt_str_read_next_line(options_line_buffer,
-                              options_line_inputfileptr) != EOF)
-    {
-      char *tok, *lineptr = gt_str_get(options_line_buffer);
-      bool parse_ii = false, parse_qii = false, parse_minid = false,
-           parse_history = false;;
-
-      while (!had_err && (tok = strsep(&lineptr," ")) != NULL)
-      {
-        if (parse_minid || parse_history)
-        {
-          GtWord readgtword;
-
-          if (sscanf(tok,GT_WD,&readgtword) != 1 ||
-              readgtword < 0  || (parse_minid && readgtword > 99) ||
-              (parse_history && readgtword > 64))
-          {
-            gt_error_set(err,"cannot parse argument for option -%s from "
-                             "first line of file %s",
-                             parse_minid ? "minidentity"
-                                         : "history",
-                             semi->matchfilename);
-            had_err = -1;
-          }
-          if (parse_minid)
-          {
-            semi->errorpercentage = 100 - readgtword;
-            parse_minid = false;
-          } else
-          {
-            semi->history_size = (GtUword) readgtword;
-            parse_history = false;
-          }
-          continue;
-        }
-        if (parse_ii)
-        {
-          gt_str_set(semi->ii, tok);
-          parse_ii = false;
-          continue;
-        }
-        if (parse_qii)
-        {
-          gt_str_set(semi->qii, tok);
-          parse_qii = false;
-          continue;
-        }
-        if (strcmp(tok, "-ii") == 0)
-        {
-          parse_ii = true;
-          continue;
-        }
-        if (strcmp(tok, "-qii") == 0)
-        {
-          parse_qii = true;
-          continue;
-        }
-        if (strcmp(tok, "-minidentity") == 0)
-        {
-          parse_minid = true;
-          continue;
-        }
-        if (strcmp(tok, "-history") == 0)
-        {
-          parse_history = true;
-          continue;
-        }
-        if (strcmp(tok, "-mirror") == 0)
-        {
-          semi->mirror = true; /* found -mirror option */
-        }
-        if (strcmp(tok, "-bias-parameters") == 0)
-        {
-          semi->bias_parameters = true;
-        }
-      }
-      if (!had_err)
-      {
-        if (gt_str_length(semi->ii) == 0UL) {
-          gt_error_set(err, "need output of option string "
-                            "(run gt seed_extend with -v or -verify)");
-          had_err = -1;
-        }
-        if (semi->errorpercentage == 0)
-        {
-          gt_error_set(err,"missing option -minidentity in first line of file "
-                           "%s",semi->matchfilename);
-          had_err = -1;
-        }
-      }
-    } else
-    {
-      gt_error_set(err, "file %s is empty", semi->matchfilename);
-      had_err = -1;
-    }
-    gt_str_delete(options_line_buffer);
-  }
-  if (options_line_inputfileptr != NULL)
-  {
-    fclose(options_line_inputfileptr);
-  }
-  /* Load encseqs */
-  if (!had_err) {
-    encseq_loader = gt_encseq_loader_new();
-    gt_encseq_loader_enable_autosupport(encseq_loader);
-    semi->aencseq
-      = gt_encseq_loader_load(encseq_loader, gt_str_get(semi->ii), err);
-    if (semi->aencseq == NULL) {
-      had_err = -1;
-    }
-  }
-  if (!had_err) {
-    if (gt_str_length(semi->qii) != 0) {
-      semi->bencseq
-        = gt_encseq_loader_load(encseq_loader, gt_str_get(semi->qii), err);
-    } else {
-      semi->bencseq = gt_encseq_ref(semi->aencseq);
-    }
-    if (semi->bencseq == NULL) {
-      had_err = -1;
-      gt_encseq_delete(semi->aencseq);
-    }
-  }
-  gt_encseq_loader_delete(encseq_loader);
-  if (!had_err)
-  {
-    gt_assert(semi->bencseq != NULL);
-    semi->query_totallength = gt_encseq_total_length(semi->bencseq);
-    semi->line_buffer = gt_str_new();
-    semi->linenum = 0;
-    semi->inputfileptr = fopen(semi->matchfilename, "rb");
-    if (semi->inputfileptr == NULL) {
-      gt_error_set(err, "file %s does not exist", semi->matchfilename);
-      had_err = true;
-    }
-  }
-  if (had_err)
-  {
-    gt_seedextend_match_iterator_delete(semi);
-    return NULL;
-  }
-  return semi;
 }
 
 static void gt_show_seed_extend_plain(LinspaceManagement *linspace_spacemanager,
@@ -425,134 +214,6 @@ static void gt_show_seed_extend_encseq(GtQuerymatch *querymatchptr,
   {
     gt_querymatch_prettyprint(querymatchptr);
   }
-}
-
-int gt_seedextend_match_iterator_next(GtSeedextendMatchIterator *semi,
-                                      GtError *err)
-{
-  int had_err = 0;
-  bool selfmatch = semi->aencseq == semi->bencseq ? true : false;
-
-  while (!had_err)
-  {
-    const char *line_ptr;
-
-    semi->linenum++;
-    if (gt_str_read_next_line(semi->line_buffer, semi->inputfileptr) == EOF)
-    {
-      gt_str_reset(semi->line_buffer);
-      return 0;
-    }
-    line_ptr = gt_str_get(semi->line_buffer);
-    /* ignore comment lines; but print seeds if -seed-display is set */
-    gt_assert(line_ptr != NULL);
-    if (line_ptr[0] != '\n')
-    {
-      if (line_ptr[0] == '#')
-      {
-        const char *seedwordptr = strstr(line_ptr, "seed:");
-
-        if (seedwordptr != NULL)
-        {
-          int num = sscanf(seedwordptr + sizeof ("seed:"),
-                           GT_WU " " GT_WU " " GT_WU,
-                           &semi->seedpos1,&semi->seedpos2,
-                           &semi->seedlen);
-          if (num != 3)
-          {
-            gt_error_set(err, "file %s, line %" PRIu64
-                         ": cannot parse 'seed:'-line \"%s\"",
-                         semi->matchfilename,
-                         semi->linenum,line_ptr);
-            had_err = -1;
-          } else
-          {
-            semi->has_seedline = true;
-          }
-        }
-      } else
-      {
-        if (gt_querymatch_read_line(semi->querymatchptr,
-                                    line_ptr,
-                                    selfmatch,
-                                    semi->query_totallength,
-                                    semi->aencseq))
-        {
-          gt_str_reset(semi->line_buffer);
-          return 1;
-        }
-      }
-    }
-    gt_str_reset(semi->line_buffer);
-  }
-  return had_err;
-}
-
-const GtEncseq *gt_seedextend_match_iterator_aencseq(
-                        const GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->aencseq;
-}
-
-const GtEncseq *gt_seedextend_match_iterator_bencseq(
-                        const GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->bencseq;
-}
-
-GtUword gt_seedextend_match_iterator_history_size(
-                        const GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->history_size;
-}
-
-GtUword gt_seedextend_match_iterator_errorpercentage(
-                        const GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->errorpercentage;
-}
-
-bool gt_seedextend_match_iterator_bias_parameters(
-                        const GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->bias_parameters;
-}
-
-bool gt_seedextend_match_iterator_has_seedline(
-                        const GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->has_seedline;
-}
-
-GtQuerymatch *gt_seedextend_match_iterator_querymatch_ptr(
-                        GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->querymatchptr;
-}
-
-GtUword gt_seedextend_match_iterator_seedlen(GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->seedlen;
-}
-
-GtUword gt_seedextend_match_iterator_seedpos1(GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->seedpos1;
-}
-
-GtUword gt_seedextend_match_iterator_seedpos2(GtSeedextendMatchIterator *semi)
-{
-  gt_assert(semi != NULL);
-  return semi->seedpos2;
 }
 
 static int gt_show_seedext_runner(GT_UNUSED int argc,
